quickstart/KalevalaGPT/finetune/requirements.txt
accelerate
peft
bitsandbytes
transformers
trl
datasets
pyyaml
mistral_common
sentencepiece